requirements.txt
# Core RL libraries
stable-baselines3>=2.0.0
gymnasium>=0.28.0
mujoco>=2.3.0
dm-control>=1.0.0

# Experiment tracking
tensorboard>=2.13.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
joblib>=1.3.0
rich>=13.0.0
